ADD query_capture default Test
